+++ src/app/navigation/navigation.module.ts
@@ -80,6 +80,7 @@ import { CheckboxStyleComponent } from '../styles/checkbox-style/checkbox-style.
 import { LabJSComponent } from '../styles/lab-js/lab-js.component';
 import { HtmlTagStyleComponent } from '../styles/html-tag-style/html-tag-style.component';
 import { ShepherdJsStyleComponent } from '../styles/shepherd-js-style/shepherd-js-style.component';
+import { EntryRecordDeleteStyleComponent } from '../styles/entry-record-delete-style/entry-record-delete-style.component';
 PlotlyModule.plotlyjs = PlotlyJS;
 
 const routes: Routes = [
@@ -192,7 +193,8 @@ const routes: Routes = [
         CheckboxStyleComponent,
         LabJSComponent,
         HtmlTagStyleComponent,
-        ShepherdJsStyleComponent
+        ShepherdJsStyleComponent,
+        EntryRecordDeleteStyleComponent
     ]
 })
 export class NavigationPageModule {